+++ selenium_class/distribuicao_federal.py
@@ -26,10 +26,13 @@ class Distribuicao_federal:
         self._definicao = pName
         
         self._error._getcoll('error')
-        self._save = '/opt/certidao/download/'
+        self._save = '/opt/certidao/download/distribuicao_federal'
         try:
-            shutil.rmtree(self._save)
-            os.makedirs(f'{self._save}')
+            if os.path.isdir(f'{self._save}') is False:
+                os.makedirs(f'{self._save}')
+            else:
+                shutil.rmtree(self._save)
+                os.makedirs(f'{self._save}')
         except:
             pass
         self._pasta = self._data['path']
