+++ selenium_class/federal.py
@@ -22,10 +22,13 @@ class Federal:
         self._error = pError
         
         self._error._getcoll('error')
-        self._save = '/opt/certidao/download/'
+        self._save = '/opt/certidao/download/federal'
         try:
-            shutil.rmtree(self._save)
-            os.makedirs(f'{self._save}')
+            if os.path.isdir(f'{self._save}') is False:
+                os.makedirs(f'{self._save}')
+            else:
+                shutil.rmtree(self._save)
+                os.makedirs(f'{self._save}')
         except:
             pass
         self._pasta = self._data['path']
