+++ selenium_class/protesto2.py
@@ -24,10 +24,13 @@ class Protesto2:
         self._error = pError
         self._captcha = pCaptcha
         self._error._getcoll('error')
-        self._save = '/opt/certidao/download/'
+        self._save = '/opt/certidao/download/protesto2'
         try:
-            shutil.rmtree(self._save)
-            os.makedirs(f'{self._save}')
+            if os.path.isdir(f'{self._save}') is False:
+                os.makedirs(f'{self._save}')
+            else:
+                shutil.rmtree(self._save)
+                os.makedirs(f'{self._save}')
         except:
             pass
         self._pasta = self._data['path']
