+++ selenium_class/trf.py
@@ -25,10 +25,13 @@ class Trf:
         self._captcha = pCaptcha
         self._error._getcoll('error')
         self._cont = 0
-        self._save = '/opt/certidao/download/'
+        self._save = '/opt/certidao/download/trf'
         try:
-            shutil.rmtree(self._save)
-            os.makedirs(f'{self._save}')
+            if os.path.isdir(f'{self._save}') is False:
+                os.makedirs(f'{self._save}')
+            else:
+                shutil.rmtree(self._save)
+                os.makedirs(f'{self._save}')
         except:
             pass
         self._pasta = self._data['path']
